-plygonet: Merged into one file and did a standardization pass. [Ryan Holtz]
